deleta hook useInProgressRecipeHook e cria logica interna no componente FoodDetails
